Use sys.executable in lldb-dotest
Without that, dotest.py would be executed with the default python
interpreter, which may not be the same one that lldb is built with.

This still requires the user do know which python interpreter to use
when running lldb-dotest, but now he is at least able to choose it, if
he knows which one to use.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@354105 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 17abf8fec08f5daa0992c07c8022f81726ec90b9
apple-llvm-split-dir: lldb/