[CMake] Disable individual modules setting for now
The idea here was to pass a set of options that is relevant for LLDB to all involved projects (LLVM, Clang, Swift). However, it requires a working caches mechanism and so far I struggled to wire this up in build-script.
Most of the other settings are fine when applied locally to LLDB, but modules should not be force enabled here individually.

apple-llvm-split-commit: 99c58213f552640be6de6e007bfbb2fe18debb0c
apple-llvm-split-dir: lldb/