Fix TestPaths.py on windows
I committed an implementation of GetClangResourceDir on windows but
forgot to update this test. I merged the tests like I intended to, but I
realized that the test was actually failing. After looking into it, it
appears that FileSystem::Resolve was taking the path and setting
the FileSpec's Directory to "/path/to/lldb/lib/clang/" and the File to
"9.0.0" which isn't what we want. So I removed the resolve line from
DefaultComputeClangResourceDir.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@355648 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry-picked from commit b3fe704527ae502410fea4e189d1df3d923a0a5f)

apple-llvm-split-commit: 455bf53f4eb9bac2dfb6dfbda41c83180f3c6657
apple-llvm-split-dir: lldb/