Rework how lldb warngs about kexts that failed to load in a darwin kernel debug session.
Originally, the kext name & uuid were emitted in the middle of the
kext-loading period's.  Last week I decided to try not printing
any details about kexts that failed to load, only printing a summary
of how many failed to load.  

This time I'm print different progress characters depending on whether
the kext loaded or not ("-" for not), then at the end I will print a 
summary of how many kexts failed to load and a sorted list of the
kexts with the bundle ID and the uuid.  It's a lot more readable.

<rdar://problem/48654569> 



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@355958 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit e8e94ccea5e3ef200065dbaf60b05cab13ba8d43)

apple-llvm-split-commit: d694df378446063bc7507fb0e2b0c1bb2aa4e1f4
apple-llvm-split-dir: lldb/