[testsuite] Split Objective-C data formatter
The testcase for objective-c data formatters is very big as it checks a
bunch of stuff. This is annoying when using the lit test driver, because
it prevents us from running the different cases in parallel. As a
result, it's always one of the last few tests that complete. This patch
splits the test into multiple files that share a common base class. This
way lit can run the different tests in parallel.

Differential revision: https://reviews.llvm.org/D60300

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@357786 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 84dac1209f3249517c9ec3c7c4ae70e77f0a77a0)

apple-llvm-split-commit: 096f3c4e2db34f05eb8b4e9446c652c962d906aa
apple-llvm-split-dir: lldb/