[Docs] Generate the python reference without building all of LLDB
As discussed on the mailing list, we should be able to generate the
Python reference without building all of LLDB. To make that possible I
create a dummy python package, which is then parsed by epydoc. The
latter will complain that it couldn't import lldb, but that doesn't
matter as far as generation of the docs is concerned.

Differential revision: https://reviews.llvm.org/D61216

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@359465 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 3754bdaf8b9f00145396103718d8a8576b83f0f0
apple-llvm-split-dir: lldb/