[Swift] Refactor all module loader creation/addition to happen in GetASTContext()
This is necessary to be able to use the ClangImporter's clang::CompilerInstance
to get the default clang module cache path to use in Swift's
ParseableInterfaceModuleLoader which must be added before adding the
ClangImporter. Previously they were created and added in separate methods
(GetASTContext() for the PIML and GetClangImporter() for the ClangImporter).

apple-llvm-split-commit: ce544570962eec27434202370276d850703c49d3
apple-llvm-split-dir: lldb/