Remove SwiftASTContext::isTrivialOptionSetType. (NFC)
To serve the SwiftOptionSet data formatter, SwiftASTContext had an
elaborate mechanism in place to determine whether the type is
"trivial" or not. Nowadays, for all practical intents and purposes
quering for the existence of the rawValue child, appears to have the
same effect and was done by the data formatter anyway.

rdar://problem/50735738

apple-llvm-split-commit: 048c38cf31219ba65c41d8ee0c2e7fd0751dd278
apple-llvm-split-dir: lldb/