[Target] Generalize language-specific behavior in ThreadPlanStepThrough
Summary:
When creating a ThreadPlan to step through a trampoline, we ask the
ObjC language runtime and the CPP language runtime to come up with such a thread
plan if the dynamic loader fails to give us one. I don't see why this behavior
can't be language agnostic.

Differential Revision: https://reviews.llvm.org/D61921

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@362164 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 1bb391f3efb03ac5d87f7901f589b4dfde4edec3
apple-llvm-split-dir: lldb/