Surface Swift deserialization problems to the user.
Error messages such as .swiftmodules being produced by an older
compiler are common, but were previously not visible to and LLDB user
unless they also turned on the types log. This patch changes that by
surfacing the errors on the LLDB error stream, but only once per
module.

<rdar://problem/51686950>

apple-llvm-split-commit: 37472aa999e26387606e4368f818489d332e7377
apple-llvm-split-dir: lldb/