Set DYLD_LIBRARY_PATH to test with the just-built stdlib
With Swift in the OS, we need to ensure that the test suite runs with
the standard library we just built, instead of the one form the
operating system. We do this by telling dyld to pick the one from our
build directory by setting the `DYLD_LIBRARY_PATH`. Swift takes the same
approach in their test suite.

apple-llvm-split-commit: 3394ff9d308a90689a5561ce97d57171dc572b73
apple-llvm-split-dir: lldb/